@@ -15,28 +15,7 @@
 
 #include "vint.h"
 #include "timer.h"
-
-struct cassandra_type
-{
-    std::string cql_name;
-    size_t fixed_len;
-    std::shared_ptr<arrow::DataType> arrow_type;
-};
-
-const std::string listtype = "org.apache.cassandra.db.marshal.ListType";
-const std::string maptype = "org.apache.cassandra.db.marshal.MapType";
-const std::string settype = "org.apache.cassandra.db.marshal.SetType";
-const std::string tupletype = "org.apache.cassandra.db.marshal.TupleType"; // TODO
-
-// order matters and is used in sstable_data.ksy to determine how child cells
-// are encoded
-const std::vector<std::string> multi_cell_types{
-    listtype,
-    maptype,
-    settype};
-
-extern const std::map<std::string, struct cassandra_type> type_info;
-extern const std::set<std::string> is_multi_cell;
+#include "conversions.h"
 
 class deserialization_helper_t : public kaitai::kstruct
 {
@@ -69,7 +48,6 @@ public:
     static bool is_multi_cell(int kind, int i);
 
     int get_col_size();
-    int get_col_size(const std::string &type);
 
     static int inc();
 };
