@@ -2,21 +2,23 @@
 
 vint_t::vint_t(kaitai::kstream *ks) : kaitai::kstruct(ks)
 {
-    char first_byte = ks->read_bytes(1)[0];
-    if (first_byte >= 0)
-    {
+    int8_t first_byte = ks->read_s1();
+    if ((first_byte & 0x80) == 0) // positive
         val_ = first_byte;
-        return;
-    }
-
-    int size = __builtin_clz(~first_byte) - 24; // -24 since C++ treats it as a 4-byte integer but we only want the number of leading zeros in the last byte
-    val_ = first_byte & (0xff >> size);         // get the value of the remaining part of it
-
-    for (int i = 0; i < size; i++)
+    else
     {
-        val_ <<= 8;
-        val_ |= ks->read_bytes(1)[0] & 0xff;
+        // get number of leading set bits
+        int size = (uint8_t)first_byte == 0xff ? 8 : __builtin_clz(~first_byte) - 24; // -24 since C++ treats it as a 4-byte integer but we only want the number of leading zeros in the single byte
+        val_ = first_byte & (0xff >> size);                                           // get the value of the remaining part of it
+
+        for (int i = 0; i < size; i++)
+        {
+            char c = ks->read_u1();
+            val_ <<= 8;
+            val_ |= c & 0xff;
+        }
     }
+    DEBUG_ONLY(std::cout << "read vint " << std::hex << val_ << std::dec << '\n');
 }
 
 // TODO
